Temporary fix for seed file and database schema

+++ db/migrate/20150730192401_change_amount_on_recipe_ingredients_to_decimal.rb
@@ -1,5 +1,5 @@
 class ChangeAmountOnRecipeIngredientsToDecimal < ActiveRecord::Migration
   def change
-    change_column :recipe_ingredients, :amount, :decimal
+    change_column :recipe_ingredients, :amount, :integer
   end
 end
